@@ -1,17 +1,16 @@
-import { observable, action, decorate } from "mobx";
-import {TODAS} from "./../constants/categories.js"
+import { observable, action, decorate } from 'mobx';
+import { TODAS } from './../constants/categories.js';
 
 class CategoryState {
-    current = TODAS;
-    setCurrent(current) {
-        this.current = current;
-    }
+  current = TODAS;
+  setCurrent(current) {
+    this.current = current;
+  }
 }
 
 decorate(CategoryState, {
-    current: observable,
-    setCurrent: action
+  current: observable,
+  setCurrent: action,
 });
 
 export default new CategoryState();
-
