@@ -1,9 +1,9 @@
-import React, { Component } from 'react';
+import React from 'react';
 import styled from 'styled-components';
 import Product from './Product';
-import {observer} from 'mobx-react';
-import CategoryState from "./../../CategoryState";
-import CounterStore from "./../../ShoppingCart";
+import { observer } from 'mobx-react';
+import CategoryState from './../../CategoryState';
+import CounterStore from './../../ShoppingCart';
 
 const Container = styled.div`
   margin-top: -50px;
@@ -31,7 +31,6 @@ const List = styled.ul`
   li:hover {
     cursor: pointer;
   }
-
 `;
 
 const Grid = styled.div`
@@ -43,35 +42,49 @@ const Grid = styled.div`
   grid-gap: 1em;
   padding: 1em;
   flex: 3;
-`
-const onSelectedProductHandler = (id) => {
+`;
+const onSelectedProductHandler = id => {
   CounterStore.currentProduct = id;
-}
+};
 
-const Categories = ({categories, products}) => {
+const Categories = ({ categories, products }) => {
+  const categoryList = categories.map(c => (
+    <li
+      key={c.id}
+      href={`#${c.id}`}
+      onClick={() => CategoryState.setCurrent(c.id)}>
+      {c.name}
+    </li>
+  ));
 
-    const categoryList = categories.map(c => (
-      <li key={c.id} href={`#${c.id}`} onClick={()=>CategoryState.setCurrent(c.id)} >{c.name}</li>
-    ));
-
-    const filteredProducts = products
-      .filter((p) => p.category === CategoryState.current || CategoryState.current  === "todas")
-      .map(function(p){
-        return <Product clicked={() => onSelectedProductHandler(p.id)} key={p.id} name={p.name} price={p.price}/>
-      });
-
-    return (
-      <Container>
-          <List>
-            <li onClick={()=>CategoryState.setCurrent("todas")} className="title">Categorías</li>
-            {categoryList}
-          </List>
-          <Grid>
-            {filteredProducts}
-          </Grid>
-      </Container>
+  const filteredProducts = products
+    .filter(
+      p =>
+        p.category === CategoryState.current ||
+        CategoryState.current === 'todas'
     )
+    .map(function(p) {
+      return (
+        <Product
+          clicked={() => onSelectedProductHandler(p.id)}
+          key={p.id}
+          name={p.name}
+          price={p.price}
+        />
+      );
+    });
 
+  return (
+    <Container>
+      <List>
+        <li onClick={() => CategoryState.setCurrent('todas')} className="title">
+          Categorías
+        </li>
+        {categoryList}
+      </List>
+      <Grid>{filteredProducts}</Grid>
+    </Container>
+  );
 };
 
 export default observer(Categories);
