@@ -1,29 +1,29 @@
-export const TODAS = "todas";
-export const ALEBRIJES = "alebrijes";
-export const BOLSAS = "bolsas";
-export const CAJAS = "cajas";
-export const CRUCES = "cruces";
-export const DECORACION = "decoracion";
-export const MOLCAJETES = "molcajetes";
-export const MUNECAS = "munecas";
-export const PAPEL = "papel";
-export const PINATAS = "pinatas";
-export const TEQUILEROS = "tequileros";
-export const TORTILLEROS = "tortilleros";
-export const VIRGENES = "virgenes";
+export const TODAS = 'todas';
+export const ALEBRIJES = 'alebrijes';
+export const BOLSAS = 'bolsas';
+export const CAJAS = 'cajas';
+export const CRUCES = 'cruces';
+export const DECORACION = 'decoracion';
+export const MOLCAJETES = 'molcajetes';
+export const MUNECAS = 'munecas';
+export const PAPEL = 'papel';
+export const PINATAS = 'pinatas';
+export const TEQUILEROS = 'tequileros';
+export const TORTILLEROS = 'tortilleros';
+export const VIRGENES = 'virgenes';
 
 export const categories = [
-    {id: TODAS, name:"Todas"},
-    {id: ALEBRIJES, name: "Alebrijes"},
-    {id: BOLSAS, name: "Bolsas"},
-    {id: CAJAS, name: "Cajas Multiusos"},
-    {id: CRUCES, name: "Cruces"},
-    {id: DECORACION, name: "Decoración"},
-    {id: MOLCAJETES, name: "Molcajetes"},
-    {id: MUNECAS, name: "Muñecas Marías"},
-    {id: PAPEL, name: "Papel picado"},
-    {id: PINATAS, name: "Piñatas"},
-    {id: TEQUILEROS, name: "Tequileros"},
-    {id: TORTILLEROS, name: "Tortilleros"},
-    {id: VIRGENES, name: "Vírgenes"}
+  { id: TODAS, name: 'Todas' },
+  { id: ALEBRIJES, name: 'Alebrijes' },
+  { id: BOLSAS, name: 'Bolsas' },
+  { id: CAJAS, name: 'Cajas Multiusos' },
+  { id: CRUCES, name: 'Cruces' },
+  { id: DECORACION, name: 'Decoración' },
+  { id: MOLCAJETES, name: 'Molcajetes' },
+  { id: MUNECAS, name: 'Muñecas Marías' },
+  { id: PAPEL, name: 'Papel picado' },
+  { id: PINATAS, name: 'Piñatas' },
+  { id: TEQUILEROS, name: 'Tequileros' },
+  { id: TORTILLEROS, name: 'Tortilleros' },
+  { id: VIRGENES, name: 'Vírgenes' },
 ];
