@@ -1,15 +1,52 @@
-import {ALEBRIJES, BOLSAS, CAJAS, DECORACION,
-    MOLCAJETES, TORTILLEROS, VIRGENES} from "./categories";
+import {
+  ALEBRIJES,
+  BOLSAS,
+  CAJAS,
+  DECORACION,
+  MOLCAJETES,
+  TORTILLEROS,
+  VIRGENES,
+} from './categories';
 
 export const products = [
-    {id: 1, name: "Jarrón de porcelana china 1", price: 800, category: ALEBRIJES},
-    {id: 2, name: "Jarrón de porcelana china 2", price: 800, category: TORTILLEROS},
-    {id: 3, name: "Jarrón de porcelana china 3", price: 200, category: DECORACION},
-    {id: 4, name: "Jarrón de porcelana china 4", price: 800, category: MOLCAJETES},
-    {id: 5, name: "Jarrón de porcelana china 5", price: 800, category: BOLSAS},
-    {id: 6, name: "Jarrón de porcelana china 6", price: 800, category: CAJAS},
-    {id: 7, name: "Jarrón de porcelana china 7", price: 200, category: DECORACION},
-    {id: 8, name: "Jarrón de porcelana china 8", price: 800, category: CAJAS},
-    {id: 9, name: "Jarrón de porcelana china 9", price: 200, category: BOLSAS},
-    {id: 10, name: "Jarrón de porcelana china 10", price: 200, category: BOLSAS}
+  {
+    id: 1,
+    name: 'Jarrón de porcelana china 1',
+    price: 800,
+    category: ALEBRIJES,
+  },
+  {
+    id: 2,
+    name: 'Jarrón de porcelana china 2',
+    price: 800,
+    category: TORTILLEROS,
+  },
+  {
+    id: 3,
+    name: 'Jarrón de porcelana china 3',
+    price: 200,
+    category: DECORACION,
+  },
+  {
+    id: 4,
+    name: 'Jarrón de porcelana china 4',
+    price: 800,
+    category: MOLCAJETES,
+  },
+  { id: 5, name: 'Jarrón de porcelana china 5', price: 800, category: BOLSAS },
+  { id: 6, name: 'Jarrón de porcelana china 6', price: 800, category: CAJAS },
+  {
+    id: 7,
+    name: 'Jarrón de porcelana china 7',
+    price: 200,
+    category: DECORACION,
+  },
+  { id: 8, name: 'Jarrón de porcelana china 8', price: 800, category: CAJAS },
+  { id: 9, name: 'Jarrón de porcelana china 9', price: 200, category: BOLSAS },
+  {
+    id: 10,
+    name: 'Jarrón de porcelana china 10',
+    price: 200,
+    category: VIRGENES,
+  },
 ];
